Refactor: atom, molecules, organisms shared 폴더 아래로 이동
