+++ src/components/index.ts
@@ -1,7 +1,5 @@
-/* eslint-disable import/no-cycle */
-export * from './molecules';
-export * from './organisms';
 export * from './layout';
 export * from './dashboard';
 export * from './eventEdit';
 export * from './login';
+export * from './shared';
