+++ src/components/shared/index.ts
@@ -1,0 +1,3 @@
+export * from './molecules';
+export * from './organisms';
+export * from './atom';
