+++ src/components/shared/molecules/input/SelectPhotoboothInput.tsx
